Add auto-updater functionality for validator.
